requirements.txt
pytelegrambotapi==3.5.2
google-api-python-client==1.6.5
google_cloud_storage==1.8.0
googleads==10.0.0
beaker==1.9.0
psycopg2==2.7.4
crontab==0.22.0
pytest==3.4.2
money==1.3.0
apscheduler==3.5.1
requests-mock==1.4.0
git+git://github.com/kolyabehemoth/python-telegram-bot
git+git://github.com/kolyabehemoth/pyquibase